Customize visual-line-mode and truncation
I don't want the modifications to editing commands like C-k that
visual-line does in prog-modes so I'm disabling it globally and only
enabling it in text-modes for now.

Enable truncate-lines globally. It will not have any effect when
visual-line-mode is in use so in practice it will just change the
default from naive text wrapping to truncation.